+++ client/src/App.css
@@ -40,7 +40,7 @@
 }
 
 .section-container {
-  margin-bottom: 10rem;
+  margin-bottom: 5rem;
 }
 
 .section-head {
